@@ -1,4 +1,5 @@
 [tool]
+
 [tool.poetry]
 name = "naja"
 version = "0.1.0"
@@ -23,9 +24,7 @@ packages = [{ include = "naja" }, { include = "tests", format = "sdist" }]
 [tool.poetry.dependencies]
 python = ">=3.8.18,<4.0"
 click = "8.0.1"
-
-black = { version = "^21.5b2", optional = true }
-isort = { version = "^5.8.0", optional = true }
+setuptools = "~69.1"
 mypy = { version = "^0.900", optional = true }
 pytest = { version = "^6.2.4", optional = true }
 pytest-cov = { version = "^2.12.0", optional = true }
@@ -43,10 +42,11 @@ pre-commit = { version = "^2.12.0", optional = true }
 toml = { version = "^0.10.2", optional = true }
 bump2version = { version = "^1.0.1", optional = true }
 httpx = "^0.25.0"
-tldextract = "^3.6.0"
-hypothesis = { version = "^6.88.1", optional = true }
+tldextract = "^4.0.0"
+hypothesis = { version = "6.88.1", optional = true }
 pytest-xdist = { version = "^3.3.1", optional = true }
 pytest-sugar = { version = "^1.0.0", optional = true }
+ruff = { version = "^0.4.2", optional = true }
 
 [tool.poetry.extras]
 test = [
@@ -80,25 +80,126 @@ doc = [
 [tool.poetry.scripts]
 naja = 'naja.cli:main'
 
-[tool.black]
+
+[tool.ruff]
+# Exclude a variety of commonly ignored directories.
+exclude = [
+    ".bzr",
+    ".direnv",
+    ".eggs",
+    ".git",
+    ".git-rewrite",
+    ".hg",
+    ".mypy_cache",
+    ".nox",
+    ".pants.d",
+    ".pytype",
+    ".ruff_cache",
+    ".svn",
+    ".tox",
+    ".venv",
+    "__pypackages__",
+    "_build",
+    "buck-out",
+    "build",
+    "dist",
+    "node_modules",
+    "venv",
+    "*/migrations/*.py",
+    "staticfiles/*",
+]
 line-length = 88
-skip-string-normalization = true
-target-version = ['py36', 'py37', 'py38']
-include = '\.pyi?$'
-exclude = '''
-/(
-    \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | _build
-  | buck-out
-  | build
-  | dist
-)/
-'''
+indent-width = 4
+target-version = "py312"
+
+[tool.ruff.lint]
+select = [
+    "F",
+    "E",
+    "W",
+    "C90",
+    "I",
+    "N",
+    "UP",
+    "YTT",
+    "ANN",
+    "ASYNC",
+    "S",
+    "BLE",
+    "FBT",
+    "B",
+    "A",
+    "COM",
+    "C4",
+    "DTZ",
+    "T10",
+    "DJ",
+    "EM",
+    "EXE",
+    "FA",
+    'ISC',
+    "ICN",
+    "G",
+    'INP',
+    'PIE',
+    "T20",
+    'PYI',
+    'PT',
+    "Q",
+    "RSE",
+    "RET",
+    "SLF",
+    "SLOT",
+    "SIM",
+    "TID",
+    "TCH",
+    "INT",
+    "ARG",
+    "PTH",
+    "ERA",
+    "PD",
+    "PGH",
+    "PL",
+    "TRY",
+    "FLY",
+    "NPY",
+    "AIR",
+    "PERF",
+    "FURB",
+    "LOG",
+    "RUF",
+]
+ignore = [
+    "SIM102", # sometimes it's better to nest
+    "S314",
+    "ISC001",
+    "ANN101",
+    "ANN002",
+    "ANN003",
+    "COM812",
+    "UP035",
+    "UP006",
+    "UP007",
+]
+# Allow fix for all enabled rules (when `--fix`) is provided.
+fixable = ["ALL"]
+unfixable = []
+# Allow unused variables when underscore-prefixed.
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+docstring-code-line-length = 120
+
+[tool.ruff.lint.isort]
+force-single-line = false
+
+[tool.ruff.lint.per-file-ignores]
+"tests/**" = ["FBT001", "ANN201", "PLR0913", "S101"]
+
 
 [tool.mypy]
 check_untyped_defs = true
@@ -116,21 +217,12 @@ warn_unused_configs = true
 warn_unreachable = true
 warn_no_return = true
 
-[tool.isort]
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
-line_length = 88
-skip_gitignore = true
-# you can skip files as below
-#skip_glob = docs/conf.py
 
 [tool.pytest.ini_options]
 python_files = ["tests.py", "test_*.py"]
 python_functions = ["test_*", "it_*", "its_*"]
 testpaths = ["tests"]
+
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
