@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-from typing import Set
-from urllib.robotparser import RequestRate
+from typing import TYPE_CHECKING, Protocol
 
-from typing_extensions import Protocol
+if TYPE_CHECKING:
+    from urllib.robotparser import RequestRate
 
-__all__ = ["Filterer", "Parser", "RateLimiter", "Url"]
+__all__ = ["Parser", "RateLimiter", "Url"]
 
 
 class Url(Protocol):
@@ -38,48 +38,24 @@ class Url(Protocol):
     def filetype(self) -> str: ...
 
 
-class Filterer(Protocol):
-    """
-    Filters URLs found by ensuring it passes the requirements to be processed.
-    """
-
-    def process(self, url: Url) -> Url | None:
-        """
-        Filters a URL by ensuring it passes the requirements to be processed. Accepts a base URL and a URL to filter.
-
-        :param base: A string representing the base URL that the filtered URL must start with.
-        :type base: `str`
-
-        :param url: An optional string representing the URL to filter.
-        :type url: `str | None`
-
-        :return: Either a string representing the filtered URL or None if the provided URL must be ignored.
-        :rtype: `str | None`
-        """
-        ...
-
-
 class Parser(Protocol):
-    """
-    Parses the content of a file (webpages, or sitemaps, for example) to extract the links of interest (which are stored in the `found_links` attribute as instances of `Url`).
+    """Parses the content of a file (webpages, or sitemaps, for example) to extract
+    the links of interest (which are stored in the `found_links`
+    attribute as instances of `Url`).
 
-    :param base: The base URL of the website
-    :type base: `str`
-
-    :param url_filter: A Filterer instance that filters the URLs found
-    :type url_filter: `Filterer`
+    Args:
+        base (str): The base URL to use to resolve relative URLs
     """
 
-    found_links: Set[Url]
+    found_links: set[Url]
 
-    def __init__(self, base: str, *args, **kwargs): ...
+    def __init__(self, base: str, *args, **kwargs) -> None: ...
 
     def parse_content(self, text: str) -> None:
-        """
-        Process the content of a website and update the `found_links` attribute
+        """Process the content of a website and update the `found_links` attribute
 
-        :param text: The content to parse
-        :type base: `str`
+        Args:
+            text (str): The content of the website
         """
         ...
 
@@ -89,7 +65,8 @@ class Parser(Protocol):
 
 class RateLimiter(Protocol):
     """
-    Limits the amount of concurrent network requests to certain websites in order to avoid bans and throttling.
+    Limits the amount of concurrent network requests to certain websites
+    in order to avoid bans and throttling.
     """
 
     def configure(
@@ -99,30 +76,27 @@ class RateLimiter(Protocol):
         crawl_delay: str | None = None,
         request_rate: RequestRate | None = None,
     ) -> None:
-        """
-        Configures the rate limiter to respect the rules defined by the domain with the given parameters. In the case of a craw delay, the craw delay is ignored.
+        """Configures the rate limiter to respect the rules defined by the
+        domain with the given parameters.
 
-        :param domain: A string representing the domain to crawl. Defaults to None.
-        :type domain: `str`
+        In the case of a craw delay,
+        the craw delay is ignored.
 
-        :param request_rate: A RequestRate object representing the rate at which to make requests. Defaults to None.
-        :type request_rate: `RequestRate | None`
-
-        :param crawl_delay: A string representing the delay between each crawl in the format "<number><unit>" (as of the format used by `urllib.robotparser`). Defaults to None.
-        :type crawl_delay: `str | None`
-
-        :return: `None`
+        Args:
+            domain (str, optional): A string representing the domain to crawl.
+            Defaults to None.
+            request_rate (RequestRate, optional): A RequestRate object representing
+            the rate at which to make requests. Defaults to None.
+            crawl_delay (str, optional): A string representing the delay between each
+            crawl in the format "<number><unit>" (as of the format used by
+            `urllib.robotparser`). Defaults to None.
         """
         ...
 
     async def limit(self, url: str) -> None:
-        """
-        Asynchronously limits the specified URL.
+        """Asynchronously limits the specified URL.
 
-        :param url: The URL to limit.
-        :type url: `str`
-
-        :return: `None`
-        :rtype: `None`
+        Args:
+            url (str): The URL to limit
         """
         ...
