@@ -5,7 +5,7 @@ import click
 
 
 @click.command()
-def main(args=None):
+def main(args=None) -> int:
     """Console script for naja."""
     click.echo("Replace this message by putting your code into " "naja.cli.main")
     click.echo("See click documentation at https://click.palletsprojects.com/")
