@@ -12,7 +12,7 @@ class Error(Exception):
         super().__init__(message, *args, **kwargs)
         self.message = message or self.default_message
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f"{self.__class__.__name__}({self.message})"
 
 
