@@ -3,12 +3,15 @@ from __future__ import annotations
 import asyncio
 import time
 from functools import partial
-from typing import Callable
-from urllib.robotparser import RequestRate
+from typing import TYPE_CHECKING
 
 import tldextract
 
 from . import errors, protocols
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
+    from urllib.robotparser import RequestRate
 
 __all__ = [
     "StaticRateLimiter",
@@ -19,17 +22,17 @@ __all__ = [
 
 
 class StaticRateLimiter:
-    """
-    Static rate limiter that limits the number of requests per second by waiting for a specified amount of time between requests
+    """Limit the number of requests per second by waiting for a
+    specified amount of time between requests
 
-    :param time_in_seconds: The amount of time (in seconds) to wait between requests
-    :type time_in_seconds: `float`
+    Args:
+        time_in_seconds (float): The amount of time to wait between requests
     """
 
     def __init__(self, time_in_seconds: float) -> None:
         self.time = time_in_seconds
 
-    async def limit(self, *args, **kwargs) -> None:
+    async def limit(self, url) -> None:  # noqa: ANN001, ARG002
         """Limit by wainting for the specified amount of time"""
         await asyncio.sleep(self.time)
 
@@ -38,7 +41,6 @@ class StaticRateLimiter:
         *,
         crawl_delay: str | None = None,
         request_rate: RequestRate | None = None,
-        **kwargs,
     ) -> None:
         if crawl_delay is not None:
             new_request_delay = float(crawl_delay)
@@ -46,9 +48,9 @@ class StaticRateLimiter:
             new_request_delay = request_rate.seconds / request_rate.requests
 
         if new_request_delay < 0:
-            raise errors.InvalidConfigurationError(
-                f"The new request delay must be greater than 0 (got {new_request_delay})."
-            )
+            msg = "The new request delay must be greater "
+            "than 0 (got {new_request_delay})."
+            raise errors.InvalidConfigurationError(msg)
 
         # Use the greater of the two in order to respect all the domains
         if new_request_delay > self.time:
@@ -57,35 +59,36 @@ class StaticRateLimiter:
 
 class NoLimitRateLimiter:
     """
-    A limiter that does not limit the requests. Keep in mind that sending a lot of requests per second can result in throttling or even bans.
+    A limiter that does not limit the requests. Keep in mind that sending a
+    lot of requests per second can result in throttling or even bans.
     """
 
-    async def limit(self, *args, **kwargs) -> None:
+    async def limit(self) -> None:
         """
         Asynchronously sleeps for 0 seconds.
         """
         await asyncio.sleep(0)
 
-    def configure(self, *args, **kwargs) -> None:
+    def configure(self) -> None:
         """
         Does nothing
         """
-        pass
 
 
 class TokenBucketRateLimiter:
-    """
-    A rate limiter that limits the requests by using the token bucket algorithm
+    """Limit the requests by using the token bucket algorithm
 
-    :param tokens_per_second: The number of tokens per second
-    :type tokens_per_second: `float`
+    Args:
+        tokens_per_second (float): The amount of tokens to add to the bucket
+        per second
     """
 
     __slots__ = ("_tokens_per_second", "_tokens", "_last_refresh_time")
 
-    def __init__(self, tokens_per_second: float):
+    def __init__(self, tokens_per_second: float) -> None:
         if tokens_per_second <= 0:
-            raise ValueError("tokens_per_second must be greater than 0")
+            msg = "tokens_per_second must be greater than 0"
+            raise ValueError(msg)
 
         self._tokens_per_second = tokens_per_second
         self._tokens = 0.0
@@ -93,7 +96,8 @@ class TokenBucketRateLimiter:
 
     def _refresh_tokens(self) -> None:
         """
-        Refreshes the tokens in the bucket based on the time elapsed since the last refresh
+        Refreshes the tokens in the bucket based on the time elapsed since the
+        last refresh
         """
         current_time = time.time()
         time_elapsed = current_time - self._last_refresh_time
@@ -102,14 +106,14 @@ class TokenBucketRateLimiter:
         self._last_refresh_time = current_time
 
     def consume(self, tokens: int = 1) -> bool:
-        """
-        Check if the given number of tokens can be consumed and decrease the number of available tokens if possible.
+        """Check if the given number of tokens can be consumed and decrease the
+        number of available tokens if possible.
 
-        :param tokens: An integer representing the number of tokens to consume. Default is 1.
-        :type tokens: `int`
+        Args:
+            tokens (int, optional): The number of tokens to consume. Default is 1.
 
-        :return: A boolean value indicating whether the tokens were successfully consumed or not.
-        :rtype: `bool`
+        Returns:
+            bool: `True` if the tokens were consumed, `False` otherwise
         """
         self._refresh_tokens()
         if self._tokens >= tokens:
@@ -117,20 +121,20 @@ class TokenBucketRateLimiter:
             return True
         return False
 
-    async def limit(self, *args, **kwargs) -> None:
+    async def limit(self) -> None:
         while not self.consume(1):
             pass
 
     @property
-    def tokens(self):
+    def tokens(self) -> float:
         return self._tokens
 
     @property
-    def tokens_per_second(self):
+    def tokens_per_second(self) -> float:
         return self._tokens_per_second
 
     @property
-    def last_refresh_time(self):
+    def last_refresh_time(self) -> float:
         return self._last_refresh_time
 
     def configure(
@@ -138,20 +142,19 @@ class TokenBucketRateLimiter:
         *,
         crawl_delay: str | None = None,
         request_rate: RequestRate | None = None,
-        **kwargs,
     ) -> None:
-        """
-        Configures the rate at which requests are made to a domain by setting the tokens per second.
+        """Configures the rate at which requests are made to a domain by setting the
+        tokens per second.
 
-        :param crawl_delay: The amount of time (in seconds) to wait between requests
-        :type crawl_delay: `str`
+        Args:
+            crawl_delay (str, optional): The amount of time (in seconds) to wait between
+            requests. Defaults to None.
+            request_rate (RequestRate, optional): The rate at which requests are made to
+            a domain. Defaults to None.
 
-        :param request_rate: The rate at which requests are made to a domain
-        :type request_rate: `RequestRate`
-
-        :raises `InvalidConfigurationError`: If the new computed token rate is less than or equal to 0.
-
-        :return: `None`
+        Raises:
+            errors.InvalidConfigurationError: If the new computed token rate is less
+            than or equal to 0.
         """
         if crawl_delay is not None:
             new_token_rate = 1 / int(crawl_delay)
@@ -161,55 +164,50 @@ class TokenBucketRateLimiter:
             return
 
         if new_token_rate < 0:
-            raise errors.InvalidConfigurationError(
-                f"The new token rate must be greater than 0 (got {new_token_rate})."
-            )
+            msg = f"The new token rate must be greater than 0 (got {new_token_rate})."
+            raise errors.InvalidConfigurationError(msg)
 
-        # Update the tokens per second to be the smallest of the two, in order to respect every domain's rules
         if new_token_rate < self._tokens_per_second:
             self._tokens_per_second = new_token_rate
 
 
 class PerDomainRateLimiter:
-    """
-    Rate limiter that limits the number of requests per domain using its especified limiter instance if given, otherwise uses the default limiter
-
-    :param limiter_factory: A callable that creates a limiter instance (defaults to `StaticRateLimiter` with a 1 second delay between requests)
-    :type limiter_factory: `Callable[[], protocols.BaseLimiter]`
+    """Limit the number of requests per domain using its especified
+    limiter instance if given, otherwise uses the default limiter
     """
 
     __slots__ = {"default_limiter_factory", "_domain_to_limiter"}
 
+    DEFAULT_LIMITER_FACTORY: Callable[[], protocols.RateLimiter] = partial(
+        StaticRateLimiter, time_in_seconds=1  # type: ignore[assignment]
+    )
+
     def __init__(
         self,
-        limiter_factory: Callable[[], protocols.RateLimiter] = partial(
-            StaticRateLimiter, time_in_seconds=1
-        ),
+        limiter_factory: Callable[[], protocols.RateLimiter] | None = None,
     ) -> None:
-        self.default_limiter_factory = limiter_factory
+        self.default_limiter_factory = limiter_factory or self.DEFAULT_LIMITER_FACTORY
         self._domain_to_limiter: dict[str, protocols.RateLimiter] = {}
 
-    async def limit(self, url: str, *args, **kwargs) -> None:
-        """Limit by waiting for the limiting of the limiter instance corresponding to the domain of the given url"""
+    async def limit(self, url: str) -> None:
+        """Limit by waiting for the limiting of the limiter instance corresponding to
+        the domain of the given url"""
         await self._domain_to_limiter.get(
             self.extract_domain(url), self.default_limiter_factory()
-        ).limit(url, *args, **kwargs)
+        ).limit(url)
 
     def add_domain(
         self, url: str, limiter: protocols.RateLimiter | None = None
     ) -> None:
-        """
-        Adds a new domain to the limited domains with an optional rate limiter.
+        """Adds a new domain to the limited domains with an optional rate limiter.
 
-        :param url: A string representing the URL to extract the domain from.
-        :type url: `str`
+        Args:
+            url (str): A string representing the URL to extract the domain from.
+            limiter (protocols.RateLimiter, optional): An optional `RateLimiter`
+            instance used to limit the rate of requests to the domain. Defaults to None.
 
-        :param limiter: An optional `BaseLimiter` instance used to limit the rate of requests to the domain.
-        :type limiter: `BaseLimiter | None`
-
-        :raises: `InvalidUrlError` if the given URL does not contain a valid domain.
-
-        :return: `None`
+        Raises:
+            errors.InvalidUrlError: If the given URL does not contain a valid domain.
         """
         domain = self.extract_domain(url)
         if domain == "":
@@ -219,14 +217,10 @@ class PerDomainRateLimiter:
 
     @staticmethod
     def extract_domain(url: str) -> str:
-        """
-        Extracts the domain from a given URL.
+        """Extracts the domain from a given URL.
 
-        :param url: A string representing a URL.
-        :type url: `str`
-
-        :return: A string representing the domain name extracted from the URL.
-        :rtype: `str`
+        Returns:
+            str: A string representing the domain name extracted from the URL.
         """
         return tldextract.extract(url).domain
 
@@ -238,23 +232,21 @@ class PerDomainRateLimiter:
         request_rate: RequestRate | None = None,
         **kwargs,
     ) -> None:
+        """Configures the rate at which requests are made to a domain by defining its
+        corresponding limiter.
+
+        Args:
+            domain (str): A string representing the domain name to configure
+            the limiter for.
+            crawl_delay (str, optional): The amount of time (in seconds) to wait
+            between requests. Defaults to None.
+            request_rate (RequestRate, optional): The rate at which requests are made
+            to a domain. Defaults to None.
+
+        Raises:
+            errors.InvalidConfigurationError: If the new computed token rate is less
+            than or equal to 0.
         """
-        Configures the rate at which requests are made to a domain by configurering its corresponding limiter.
-
-        :param domain: A string representing the domain name to configure the limiter for.
-        :type domain: `str`
-
-        :param crawl_delay: The amount of time (in seconds) to wait between requests
-        :type crawl_delay: `str`
-
-        :param request_rate: The rate at which requests are made to a domain
-        :type request_rate: `RequestRate`
-
-        :raises `InvalidConfigurationError`: If domain is not provided, or neither crawl_delay nor request_rate is provided or if the new computed token rate is less than or equal to 0.
-        """
-        if domain is None:
-            raise errors.InvalidConfigurationError("A domain must be provided.")
-
         if domain not in self._domain_to_limiter:
             self.add_domain(domain)
 
